+++ core/database_manager.py
@@ -16,7 +16,6 @@ class DatabaseManager:
         self.engine = create_engine(f'sqlite:///{db_path}', echo=True)
         self.session = sessionmaker(autocommit=False, autoflush=False, bind=self.engine)()
 
-
     def initialize_tables(self):
         # Создаем таблицы, если они еще не существуют
         Base.metadata.create_all(self.engine)
@@ -197,19 +196,129 @@ class DatabaseManager:
         finally:
             self.session.close()
 
-    def save_poster_blob(self, title_id, poster_blob):
+    def process_titles(self, title_data):
+        title = title_data
+        # Сохранение данных в базу данных через DatabaseManager
         try:
-            # Проверяем, есть ли уже постер для данного тайтла, добавляем новый, если нет
-            new_poster = Poster(
-                title_id=title_id,
-                poster_blob=poster_blob,
-                last_updated=datetime.utcnow()
-            )
-            self.session.add(new_poster)
+            # Сохранение тайтла
+            if isinstance(title, dict):
+                title_data = {
+                    'title_id': title.get('id'),
+                    'code': title.get('code'),
+                    'name_ru': title.get('names', {}).get('ru'),
+                    'name_en': title.get('names', {}).get('en'),
+                    'alternative_name': title.get('names', {}).get('alternative'),
+                    'franchises': json.dumps(title.get('franchises', [])),
+                    'announce': title.get('announce'),
+                    'status_string': title.get('status', {}).get('string'),
+                    'status_code': title.get('status', {}).get('code'),
+                    'poster_path_small': title.get('posters', {}).get('small', {}).get('url'),
+                    'poster_path_medium': title.get('posters', {}).get('medium', {}).get('url'),
+                    'poster_path_original': title.get('posters', {}).get('original', {}).get('url'),
+                    'updated': title.get('updated'),
+                    'last_change': title.get('last_change'),
+                    'type_full_string': title.get('type', {}).get('full_string'),
+                    'type_code': title.get('type', {}).get('code'),
+                    'type_string': title.get('type', {}).get('string'),
+                    'type_episodes': title.get('type', {}).get('episodes'),
+                    'type_length': title.get('type', {}).get('length'),
+                    'genres': json.dumps(title.get('genres', [])),
+                    'team_voice': json.dumps(title.get('team', {}).get('voice', [])),
+                    'team_translator': json.dumps(title.get('team', {}).get('translator', [])),
+                    'team_timing': json.dumps(title.get('team', {}).get('timing', [])),
+                    'season_string': title.get('season', {}).get('string'),
+                    'season_code': title.get('season', {}).get('code'),
+                    'season_year': title.get('season', {}).get('year'),
+                    'season_week_day': title.get('season', {}).get('week_day'),
+                    'description': title.get('description'),
+                    'in_favorites': title.get('in_favorites'),
+                    'blocked_copyrights': title.get('blocked', {}).get('copyrights'),
+                    'blocked_geoip': title.get('blocked', {}).get('geoip'),
+                    'blocked_geoip_list': json.dumps(title.get('blocked', {}).get('geoip_list', [])),
+                    'last_updated': datetime.utcnow()  # Использование метода utcnow()
+                }
+
+                self.save_title(title_data)
+        except Exception as e:
+            self.logger.error(f"Failed to save title to database: {e}")
+
+    def process_episodes(self, title_data):
+
+        for episode in title_data.get("player", {}).get("list", {}).values():
+            if not isinstance(episode, dict):
+                self.logger.error(f"Invalid type for episode. Expected dict, got {type(episode)}")
+                continue
+
+            if "hls" in episode:
+                try:
+                    episode_data = {
+                        'episode_id': episode.get('id'),
+                        'title_id': title_data.get('id'),
+                        'episode_number': episode.get('episode'),
+                        'name': episode.get('name', f'Серия {episode.get("episode")}'),
+                        'uuid': episode.get('uuid'),
+                        'created_timestamp': episode.get('created_timestamp'),
+                        'hls_fhd': episode.get('hls', {}).get('fhd'),
+                        'hls_hd': episode.get('hls', {}).get('hd'),
+                        'hls_sd': episode.get('hls', {}).get('sd'),
+                        'preview_path': episode.get('preview'),
+                        'skips_opening': json.dumps(episode.get('skips', {}).get('opening', [])),
+                        'skips_ending': json.dumps(episode.get('skips', {}).get('ending', []))
+                    }
+                    self.save_episode(episode_data)
+                except Exception as e:
+                    self.logger.error(f"Failed to save episode to database: {e}")
+
+    def process_torrents(self, title_data):
+        if "torrents" in title_data and "list" in title_data["torrents"]:
+            for torrent in title_data["torrents"]["list"]:
+                url = torrent.get("url")
+                if url:
+                    try:
+                        torrent_data = {
+                            'torrent_id': torrent.get('torrent_id'),
+                            'title_id': title_data.get('id'),
+                            'episodes_range': torrent.get('episodes', {}).get('string', 'Неизвестный диапазон'),
+                            'quality': torrent.get('quality', {}).get('string', 'Качество не указано'),
+                            'quality_type': torrent.get('quality', {}).get('type'),
+                            'resolution': torrent.get('quality', {}).get('resolution'),
+                            'encoder': torrent.get('quality', {}).get('encoder'),
+                            'leechers': torrent.get('leechers'),
+                            'seeders': torrent.get('seeders'),
+                            'downloads': torrent.get('downloads'),
+                            'total_size': torrent.get('total_size'),
+                            'size_string': torrent.get('size_string'),
+                            'url': torrent.get('url'),
+                            'magnet_link': torrent.get('magnet'),
+                            'uploaded_timestamp': torrent.get('uploaded_timestamp'),
+                            'hash': torrent.get('hash'),
+                            'torrent_metadata': torrent.get('metadata'),
+                            'raw_base64_file': torrent.get('raw_base64_file')
+                        }
+                        self.save_torrent(torrent_data)
+                    except Exception as e:
+                        self.logger.error(f"Ошибка при сохранении торрента в базе данных: {e}")
+
+    def save_poster_to_db(self, title_id, poster_blob):
+        try:
+            # Проверяем, существует ли уже постер для данного title_id
+            existing_poster = self.session.query(Poster).filter_by(title_id=title_id).first()
+            if not existing_poster:
+                # Создаем новый объект Poster и добавляем в базу
+                new_poster = Poster(title_id=title_id, poster_blob=poster_blob, last_updated=datetime.utcnow())
+
+                self.session.add(new_poster)
+            else:
+                # Обновляем существующий постер
+                existing_poster.poster_blob = poster_blob
+                existing_poster.last_updated = datetime.utcnow()
+
             self.session.commit()
+
+
         except Exception as e:
             self.session.rollback()
-            self.logger.error(f"Ошибка при сохранении блоба постера в базе данных: {e}")
+            self.logger.error(f"Ошибка при сохранении постера в базу данных: {e}")
 
     def get_poster_blob(self, title_id):
         try:
@@ -220,6 +329,19 @@ class DatabaseManager:
         except Exception as e:
             self.logger.error(f"Error fetching poster from database: {e}")
             return None
+
+    def get_titles_from_db(self, day_of_week=None, show_all=False, batch_size=None, offset=0, title_id=None):
+        """Получает список тайтлов из базы данных через DatabaseManager."""
+        try:
+            query = self.get_titles_query(day_of_week, show_all, title_id)
+            if batch_size:
+                query = query.offset(offset).limit(batch_size)
+            titles = query.all()
+            return titles
+        except Exception as e:
+            self.logger.error(f"Ошибка при загрузке тайтлов из базы данных: {e}")
+            return []
+
 
 # Модели для таблиц
 class Title(Base):
