+++ get-name.txt
@@ -54,27 +54,23 @@
 
 ## Этап 5: Подготовка Нового Интерфейса
 ### 5.1 Создание Окна для Нового Интерфейса
-- [ ] Создать новое окно интерфейса с использованием PyQt, которое подключается к базе данных для получения информации о тайтлах.
-- [ ] Добавить в интерфейс отображение тайтлов и эпизодов из базы данных.
+- [x] Создать новое окно интерфейса с использованием PyQt, которое подключается к базе данных для получения информации о тайтлах.
+- [x] Добавить в интерфейс отображение тайтлов и эпизодов из базы данных.
 
 ### 5.2 Переход к Новому Интерфейсу
-- [ ] Перенести логику отправки запросов в новый интерфейс.
-- [ ] Постепенно заменить старый интерфейс на новый, используя данные из базы.
+- [x] Перенести логику отправки запросов в новый интерфейс.
+- [x] Постепенно заменить старый интерфейс на новый, используя данные из базы.
 
+### Этап 6: Дополнительные Улучшения и Оптимизация
+### 6.1 разнести логику
+- [ ] распилить qt/app.py , возможно часть логики можно оптимизировать
+- [ ] возможно стоит вынести сохранение в бд
 
-## Этап 6: Тестирование и Оптимизация
-### 6.1 Тестирование Методов `DatabaseManager`
-- [ ] Написать юнит-тесты для каждого метода `DatabaseManager` (создание, чтение, обновление, удаление данных).
-- [ ] Тестировать работу с базой данных на корректность добавления и получения данных.
-
-### 6.2 Тестирование Интеграции
-- [ ] Тестировать взаимодействие базы данных с остальным приложением, чтобы убедиться в правильности логики сохранения и загрузки данных.
-- [ ] Тестировать новый интерфейс, чтобы проверить отображение данных и правильность работы функционала.
 
 
 ## Этап 7: Дополнительные Улучшения и Оптимизация
 ### 7.1 Оптимизация Работы с Датами
-- [ ] Добавить проверку на устаревшие данные (`last_updated`) для постеров и тайтлов, чтобы при необходимости загружать их заново.
+- [x] Добавить проверку на устаревшие данные (`last_updated`) для постеров и тайтлов, чтобы при необходимости загружать их заново.
 
 ### 7.2 Обработка Ошибок и Резервное Копирование
 - [ ] Реализовать обработку ошибок при взаимодействии с базой данных, чтобы избежать потери данных.
@@ -83,3 +79,13 @@
 ### 7.3 Улучшение Интерфейса
 - [ ] Добавить отображение истории просмотра и рейтингов тайтлов в новом интерфейсе.
 - [ ] Реализовать возможность пользователю отмечать просмотренные эпизоды и добавлять рейтинг.
+
+## Этап 9: Тестирование и Оптимизация
+### 9.1 Тестирование Методов `DatabaseManager`
+- [ ] Написать юнит-тесты для каждого метода `DatabaseManager` (создание, чтение, обновление, удаление данных).
+- [ ] Тестировать работу с базой данных на корректность добавления и получения данных.
+
+### 9.2 Тестирование Интеграции
+- [ ] Тестировать взаимодействие базы данных с остальным приложением, чтобы убедиться в правильности логики сохранения и загрузки данных.
+- [ ] Тестировать новый интерфейс, чтобы проверить отображение данных и правильность работы функционала.
+
