+++ logging.conf
@@ -29,7 +29,7 @@ args=(sys.stdout,)
 class=logging.handlers.TimedRotatingFileHandler
 level=DEBUG
 formatter=defaultFormatter
-args=('logs/debug_log.txt', 'midnight', 1, 7, 'utf-8')
+args=('logs/debug_log.txt', 'midnight', 1, 5, 'utf-8')
 
 [formatter_consoleFormatter]
 format=%(asctime)s | %(levelname)s | %(name)s.%(funcName)s | %(message)s
