+++ utils/logging_handlers.py
@@ -4,11 +4,11 @@ import os
 from logging.handlers import TimedRotatingFileHandler
 
 class CustomTimedRotatingFileHandler(TimedRotatingFileHandler):
-    def __init__(self, filename, log_dir='logs', when='midnight', interval=1, backupCount=7, encoding=None, delay=False, utc=False):
+    def __init__(self, filename, log_dir='logs', when='midnight', interval=1,  maxBytes=2000000, backupCount=5, encoding=None, delay=False, utc=False):
         self.log_dir = log_dir
         os.makedirs(self.log_dir, exist_ok=True)
         full_log_file = os.path.join(self.log_dir, filename)
-        super(CustomTimedRotatingFileHandler, self).__init__(full_log_file, when, interval, backupCount, encoding, delay, utc)
+        super(CustomTimedRotatingFileHandler, self).__init__(full_log_file, when, interval, maxBytes, backupCount, encoding, delay, utc)
 
     def getLogFileName(self, current_date):
         base_filename, file_extension = os.path.splitext(self.baseFilename)
